[macOS] Min window size: Adjust minimum address bar width (#536)
<!--
Note: This template is a reminder of our Engineering Expectations and
Definition of Done. Remove sections that don't apply to your changes.

⚠️ If you're an external contributor, please file an issue before
working on a PR. Discussing your changes beforehand will help ensure
they align with our roadmap and that your time is well spent.
-->

Task/Issue URL:
https://app.asana.com/1/137249556945/project/1209825025475019/task/1209568315204838
Tech Design URL:
https://app.asana.com/1/137249556945/project/1209825025475019/task/1209811936242195

### Description

This updates the address bar layout in macOS to allow it to be resized,
to support a minimum window size of 544 pt.

Various parts of the layout were limiting the window size and have been
adjusted:

1. The home page now shows the Dax logo inside the address bar, instead
of to the left of it.
2. The address bar has a minimum and maximum width for very large/small
window sizes.
3. The address bar can take up all of the available space (no extra
spacing on either side) at smaller window sizes.

Notes:

* The “Switch to Tab” box is hidden in the address bar at smaller window
sizes, to prioritize showing the address bar text.
* There is a fix to the trailing space for the address bar text field.
This accounts for the space taken by the AiChat, bookmark, and clear
buttons. (Without this fix, there could be extra space on the right side
of the address bar when no buttons were showing.)
* Support for resizing the window to the smallest size when the zoom or
permissions buttons are visible in the address bar will be added
separately.

Before|After
-|-
![Screenshot 2025-04-18 at 13 41
50](https://github.com/user-attachments/assets/17afb039-c71a-4b28-8dab-eb166a48f136)|![Screenshot
2025-04-18 at 13 18
31](https://github.com/user-attachments/assets/0f2861e8-c772-48a7-8aeb-991ac9d3066c)
![Screenshot 2025-04-18 at 13 42
02](https://github.com/user-attachments/assets/a88247df-52f2-45f8-a643-1d4c846f469c)|![Screenshot
2025-04-18 at 13 18
51](https://github.com/user-attachments/assets/87eeaa38-98ff-45de-b24f-7ab55e260e8e)
![Screenshot 2025-04-18 at 13 42
12](https://github.com/user-attachments/assets/eaf560e1-3e85-4657-a009-820a3e718323)|![Screenshot
2025-04-18 at 13 20
03](https://github.com/user-attachments/assets/c4ab2074-e45b-4d7e-a42d-65abffb77e7c)



### Testing Steps
1. Enable all navigation bar items (including the home button on the
left) and confirm you can resize the window to 544 pt, with the address
bar resized as expected.
2. Hide all navigation bar items and confirm you can resize the window
to 544 pt, with the address bar resized as expected.
3. Resize the window as large as possible as confirm the address bar is
resized as expected (centered with a maximum width).
4. Confirm the address bar works as expected in these scenarios:
    * On the home page.
    * While searching/entering an address.
    * While entering an address that is open in another tab.
    * While displaying a website address.
    * With the AiChat button visible/hidden in the address bar.
5. Hide all navigation bar items and confirm you can resize the window
to 544 pt.
6. Test across all supported macOS versions to confirm layout behaves as
expected as window is resized.

### Impact and Risks

Medium: The likely impact would be visual, but if the address bar is
resized poorly it could hinder the ability to read and enter text there.

—
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)