[macOS] Add overflow menu for pinned nav bar buttons (#648)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/72649045549333/task/1209813459676820?focus=true

### Description

This adds an overflow menu to the browser navigation bar. When the
browser window is resized to a small width, and there are buttons pinned
to the nav bar (e.g. the home button, or shortcuts like downloads,
passwords), the overflow menu is displayed if the buttons overflow the
available width.

#### Expected behavior

* When a window/tab is opened, a new button is added to the navigation
bar, or the window is resized, the overflow menu will appear if needed.
* When the overflow menu is needed, buttons will be moved one at a time
from the navigation bar to the overflow menu.
* When more space is available in the navigation bar, buttons will be
moved one at a time from the overflow menu back to the navigation bar.
* When a menu item in the overflow menu is clicked on, and has a related
popover, it will become visible in the navigation bar and the popover
will be displayed.

#### Primary Changes

* Adds an overflow menu button that is hidden by default.
* Tracks the pinned views, and which views are visible in the navigation
bar or moved into the overflow menu. These properties also determine the
order that views will be moved in/out of the overflow menu.
* Tracks the required width for the pinned views, and the available
space for showing the in the navigation bar (`overflowThreshold` - this
is the threshold for moving items into the overflow menu).
* When the view appears, the view is resized, the shortcut buttons in
the navigation bar change, or the zoom/permissions buttons in the
address bar change, the navigation bar is updated (moving items in or
out of the overflow menu) based on the window’s width and available
space.

### Testing Steps
<!-- Assume the reviewer is unfamiliar with this part of the app -->
1. Right-click in the navigation bar to enable the optional pinned
views:
   * Home button
   * Passwords shortcut
   * Bookmarks shortcut
   * Downloads shortcut
   * VPN shortcut
2. Open a website and zoom in or trigger permissions (e.g.
camera/microphone, pop-up window).
3. Confirm you can resize the window to a large size and all buttons are
visible in the navigation bar.
4. Confirm you can resize the window to the smallest width (544 pt) and
buttons are progressively moved into the overflow menu.
5. Open the overflow menu and select an item. Confirm it performs the
expected action (e.g. “Home” takes you to homepage; other buttons are
moved back to the navigation bar and their respective popovers are
opened).

### Impact and Risks
Low: Minor visual changes

#### What could go wrong?
Buttons could be hidden/shown unexpectedly at small window widths, or
resize the window unexpectedly

### Notes to Reviewer


https://github.com/user-attachments/assets/3aa6b066-1207-4d94-ab22-683e9fea1cc8




—
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)